Tests/LibWeb/Text/expected/Editing/editing-history-coalescence.txt

== a typed run coalesces into one undo unit ==
PASS: after typing DOM
FAIL: one undo removes the whole run DOM (expected "" got "abc")
FAIL: one undo removes the whole run selection (expected host:DIV@0->host:DIV@0 got host/0:#text@3->host/0:#text@3)
PASS: one redo restores the whole run DOM
PASS: one redo restores the whole run selection
== caret movement breaks typing coalescence ==
FAIL: undo removes only the second run DOM (expected "ab" got "abc")
FAIL: undo removes only the second run selection (expected host/0:#text@2->host/0:#text@2 got host/0:#text@3->host/0:#text@3)
FAIL: second undo removes the first run DOM (expected "" got "abc")
== programmatic selection change breaks coalescence even at the same position ==
FAIL: undo removes only the post-collapse run DOM (expected "ab" got "abc")
FAIL: second undo removes the first run DOM (expected "" got "abc")
== a backspace run coalesces and undo selects the restored text ==
PASS: after backspaces DOM
FAIL: one undo restores all deleted text DOM (expected "abcdef" got "abc")
FAIL: one undo restores all deleted text selection (expected host/0:#text@6->host/0:#text@3 got host/0:#text@3->host/0:#text@3)
PASS: one redo deletes it again DOM
PASS: one redo deletes it again selection
== a forward-delete run coalesces separately ==
PASS: after forward deletes DOM
FAIL: one undo restores both characters DOM (expected "abcdef" got "cdef")
FAIL: one undo restores both characters selection (expected host/0:#text@0->host/0:#text@1 got host/:DIV@0->host/:DIV@0)
PASS: one redo deletes them again DOM
FAIL: one redo deletes them again selection (expected host/0:#text@0->host/0:#text@0 got host/:DIV@0->host/:DIV@0)
== backspace starts a new unit after typing, but typing joins a delete unit ==
PASS: after type/backspace/type DOM
FAIL: undo reverts backspace and following text as one unit DOM (expected "abc" got "abd")
FAIL: undo reverts backspace and following text as one unit selection (expected host/0:#text@3->host/0:#text@2 got host/0:#text@3->host/0:#text@3)
FAIL: second undo reverts the typed run DOM (expected "" got "abd")
PASS: no third unit exists
== paragraph insertion joins a typing run ==
PASS: after enter and typing DOM
FAIL: one undo reverts the split and the typing DOM (expected "<div>hello</div>" got "<div>hel</div><div>xlo</div>")
FAIL: one undo reverts the split and the typing selection (expected host/0/0:#text@3->host/0/0:#text@3 got host/1/0:#text@1->host/1/0:#text@1)
PASS: enter and typing were one unit
== typing that removes a placeholder closes the open unit ==
FAIL: after typing into the new paragraph DOM (expected "ab<div>cd</div>" got "<div>ab</div><div>cd</div>")
FAIL: undo removes only the character typed after the placeholder was removed DOM (expected "ab<div>c</div>" got "<div>ab</div><div>cd</div>")
FAIL: second undo reverts the rest DOM (expected "" got "<div>ab</div><div>cd</div>")
== replacing a selection then typing extends the same unit ==
PASS: after replacing and typing DOM
FAIL: one undo restores the replaced selection DOM (expected "hello world" got "hello xyz")
FAIL: one undo restores the replaced selection selection (expected host/0:#text@6->host/0:#text@11 got host/0:#text@9->host/0:#text@9)
PASS: one redo replays the whole unit DOM
PASS: one redo replays the whole unit selection
== script execCommand insertText coalesces like typing ==
FAIL: one undo removes both insertions DOM (expected "" got "xy")
== formatting breaks a typing run ==
PASS: after bold and typing DOM
FAIL: undo removes only the typed character DOM (expected "<b>hello</b> world" got "<b>hellox</b> world")
FAIL: undo removes only the typed character selection (expected host/0/0:#text@5->host/0/0:#text@5 got host/0/0:#text@6->host/0/0:#text@6)
FAIL: second undo removes the formatting DOM (expected "hello world" got "<b>hellox</b> world")
FAIL: second undo removes the formatting selection (expected host/0:#text@0->host/0:#text@5 got host/0/0:#text@6->host/0/0:#text@6)
== script DOM mutation elsewhere does not break coalescence ==
PASS: after typing around the script mutation DOM
FAIL: one undo reverts both characters and keeps the script node DOM (expected "Z" got "abZ")
PASS: the script mutation was not recorded
